requirements.txt
# requirements.txt
fastapi
uvicorn
requests
sqlalchemy
python-dotenv